Refactor for production: add SosConfiguration
